add database

+++ src/main/java/org/example/controller/UserController.java
@@ -27,7 +27,7 @@ public class UserController {
     }
 
     @GetMapping("counter")
-    public int userCounter() {
+    public long userCounter() {
         return userService.userCounter();
     }
